content-security-policy index.php, fixed information disclosure
